Added unit tests
Work to be done to complete this application, due to unit tests showing an error with the discount logic. A problem to be resolved at a later date as this app is just to display style and experience.

+++ Checkout_Sample/Product.cs
@@ -2,9 +2,9 @@
 {
     public class Product
     {
-        public readonly decimal Price;
+        public readonly double Price;
 
-        public Product(decimal price)
+        public Product(double price)
         {
             this.Price = price;
         }
